Dynamic Programming
If the building has 7 floor, we drop the first egg in 3th.
Then we can say we divide this problem into two parts:
1. If the first egg breaks, then we have to use another one to test
the distant f from the 1th (because if the egg broke, we don't have
any egg to test it).
2. If the first egg is not, then another problem is "the number of
actions to test a 4th building (floor 4, 5, 6, and 7)".