Dynamic Programming
Big Mistake in this code.
When I write it first time, I thought I can convert the "right" part
to the "left" part then we don't have to calculate it again.
For example: 3-14 can convert to 1-11 plus the operation count * 3(the
shift variable).
However, we cannot do that because when the shift variable increased,
the operation will be more expensive, and make there has more chance
to do the different operation.

Therefore, I use the simple dp, to cache all the cost from a to b, and
solve this problem.