Insert interval
Split into two parts.
One is insert and merge
Second is merge to next.

In the insert step, we have to search the insert point.
Using the binary search can do it well.

After find the insert point n.
The next step is merge.
The program have to check the intervals[n-1] and intervals[n],
and check the intervals[n] and intervals[n+1].
We have to check intervals[n+i] and intervals[n+i+1] until the they
cannot merge together.