dynamic program problem
if knight start from the left-top, we should record the status of the
knight (current health, required health).
however, if we start from the right-down, that makes things easier.

makes a two-dimension db array to record the minimum health.

If the knight's health bigger then 0, then the minimum health is 0.
Otherwise, it should be -health + 1.